Moved delete/regen logic out of IslandManager
These methods are now stored in their respective paster class, and the SchematicPaster interface now implements deleteIsland() and regenIsland() for them to inherit.
