+++ src/main/java/com/iridium/iridiumskyblock/managers/IslandManager.java
@@ -456,26 +456,7 @@ public class IslandManager extends TeamManager<Island, User> {
         final int finalMinHeight = minHeight;
         int maxHeight = world.getMaxHeight();
 
-        String paster = IridiumSkyblock.getInstance().getConfiguration().paster.toLowerCase();
-
-        switch(paster) {
-
-            case "fawe": {
-                deleteFAWE(world, pos1, pos2, finalMinHeight, maxHeight, completableFuture, delay);
-                break;
-            }
-
-            case "worldedit": {
-                deleteWE(world, pos1, pos2, finalMinHeight, maxHeight, completableFuture, delay);
-                break;
-            }
-
-            case "internalasync": {}
-            default: {
-                deleteIA(world, pos1, pos2, finalMinHeight, maxHeight, completableFuture, delay);
-                 break;
-            }
-        }
+        IridiumSkyblock.getInstance().getSchematicManager().deleteIsland(world, pos1, pos2, finalMinHeight, maxHeight, completableFuture, delay);
     }
 
     public CompletableFuture<Void> regenerateTerrain(Island island) {
@@ -516,27 +497,7 @@ public class IslandManager extends TeamManager<Island, User> {
         final int finalMinHeight = minHeight;
         int maxHeight = world.getMaxHeight();
 
-        // TODO: update this to use an enum
-        String paster = IridiumSkyblock.getInstance().getConfiguration().paster.toLowerCase();
-
-        switch(paster) {
-
-            case "fawe": {
-                regenFAWE(world, regenWorld, pos1, pos2, pos3, pos4, finalMinHeight, maxHeight, completableFuture, delay);
-                break;
-            }
-
-            case "worldedit": {
-                regenWE(world, regenWorld, pos1, pos2, pos3, pos4, finalMinHeight, maxHeight, completableFuture, delay);
-                break;
-            }
-
-            case "internalasync": {}
-            default: {
-                regenIA(world, regenWorld, pos3, pos4, finalMinHeight, maxHeight, completableFuture, delay);
-                break;
-            }
-        }
+        IridiumSkyblock.getInstance().getSchematicManager().regenIsland(world, regenWorld, pos1, pos2, pos3, pos4, finalMinHeight, maxHeight, completableFuture, delay);
     }
 
     @Override
@@ -1148,447 +1109,5 @@ public class IslandManager extends TeamManager<Island, User> {
             }
         }
     }
-
-    // TODO: should consider chunk-batching like w/ WE
-    public void deleteIA(World world, Location pos1, Location pos2, int minHeight, int maxHeight, CompletableFuture<Void> completableFuture, int delay) {
-
-        Set<Long> modifiedChunks = new HashSet<>();
-
-        int minX = Math.min(pos1.getBlockX(), pos2.getBlockX());
-        int maxX = Math.max(pos1.getBlockX(), pos2.getBlockX());
-        int minZ = Math.min(pos1.getBlockZ(), pos2.getBlockZ());
-        int maxZ = Math.max(pos1.getBlockZ(), pos2.getBlockZ());
-
-        new BukkitRunnable() {
-
-            int y = maxHeight;
-            int x = minX;
-            int z = minZ;
-
-            @Override
-            public void run() {
-                for (int i = 0; i < IridiumSkyblock.getInstance().getConfiguration().pasterLimitPerTick; i++) {
-
-                    Block block = world.getBlockAt(x, y, z);
-
-                    while (block.isEmpty()) {
-                        if(increment()) return;
-                        block = world.getBlockAt(x, y, z);
-                    }
-
-                    if (block.getState() instanceof InventoryHolder) {
-                        ((InventoryHolder) block.getState()).getInventory().clear();
-                    }
-                    block.setType(Material.AIR, false);
-                    modifiedChunks.add(((long) (x >> 4) << 32) | ((z >> 4) & 0xFFFFFFFFL));
-
-                    increment();
-                }
-            }
-
-            public boolean increment() {
-                if (++z > maxZ) {
-                    z = minZ;
-                    if (++x > maxX) {
-                        x = minX;
-                        if (--y < minHeight) {
-                            finish();
-                            return true;
-                        }
-                    }
-                }
-                return false;
-            }
-
-            private void finish() {
-                this.cancel();
-
-                completableFuture.complete(null);
-
-                new BukkitRunnable() {
-                    @Override
-                    public void run() {
-                        if (!IridiumSkyblock.getInstance().getConfiguration().generatorType.isTerrainGenerator()) {
-                            modifiedChunks.forEach(key -> {world.refreshChunk((int) (key >> 32), key.intValue());});
-                        }
-                    }
-                }.runTaskAsynchronously(IridiumSkyblock.getInstance());
-            }
-        }.runTaskTimer(IridiumSkyblock.getInstance(), delay, 1L);
-    }
-
-    public void deleteWE(World world, Location pos1, Location pos2, int minHeight, int maxHeight, CompletableFuture<Void> completableFuture, int delay) {
-
-        com.sk89q.worldedit.world.World weWorld = BukkitAdapter.adapt(world);
-
-        try (EditSession editSession = WorldEdit.getInstance().newEditSession(weWorld)) {
-            BlockVector3 min = BlockVector3.at(pos1.getBlockX(), minHeight, pos1.getBlockZ());
-            BlockVector3 max = BlockVector3.at(pos2.getBlockX(), maxHeight, pos2.getBlockZ());
-            CuboidRegion selection = new CuboidRegion(min, max);
-
-            editSession.getChangeSet().setRecordChanges(false);
-            editSession.setReorderMode(EditSession.ReorderMode.FAST);
-            editSession.setSideEffectApplier(SideEffectSet.none());
-            ExistingBlockMask nonAirMask = new ExistingBlockMask(editSession);
-            editSession.setMask(nonAirMask);
-
-            // grab all points and remove air blocks from the selection
-            List<BlockVector3> points = new ArrayList<>();
-            for (BlockVector3 point : selection) {
-                if (nonAirMask.test(point)) {
-                    points.add(point);
-                }
-            }
-
-            points = points.reversed(); // style points
-
-            // group points into chunks
-            Map<BlockVector2, List<BlockVector3>> chunkBatches = new LinkedHashMap<>();
-            for (BlockVector3 pt : points) {
-                BlockVector2 chunkPos = BlockVector2.at(pt.getX() >> 4, pt.getZ() >> 4);
-                chunkBatches.computeIfAbsent(chunkPos, k -> new ArrayList<>()).add(pt);
-            }
-
-            // chunk list
-            List<Map.Entry<BlockVector2, List<BlockVector3>>> batches = new ArrayList<>(chunkBatches.entrySet());
-
-            new BukkitRunnable() {
-                int chunkIndex = 0;
-                int blockInsideChunkIndex = 0;
-                final int blocksPerTick = IridiumSkyblock.getInstance().getConfiguration().pasterLimitPerTick;
-                private final com.sk89q.worldedit.world.block.BlockState air =
-                        com.sk89q.worldedit.world.block.BlockTypes.AIR.getDefaultState();
-
-                @Override
-                public void run() {
-                    int processedThisTick = 0;
-
-                    while (processedThisTick < blocksPerTick) {
-
-                        if (chunkIndex >= batches.size()) {
-                            finish();
-                            return;
-                        }
-
-                        List<BlockVector3> currentChunkBlocks = batches.get(chunkIndex).getValue();
-
-                        while (blockInsideChunkIndex < currentChunkBlocks.size() && processedThisTick < blocksPerTick) {
-                            BlockVector3 point = currentChunkBlocks.get(blockInsideChunkIndex);
-
-                            try {
-                                editSession.setBlock(point, air);
-                            } catch (MaxChangedBlocksException e) {
-                                finish();
-                                return;
-                            }
-
-                            blockInsideChunkIndex++;
-                            processedThisTick++;
-                        }
-
-                        if (blockInsideChunkIndex >= currentChunkBlocks.size()) {
-                            chunkIndex++;
-                            blockInsideChunkIndex = 0;
-                            editSession.flushSession(); // deprecated, but this pushes changes to the server, so it's critical for performance
-                        }
-                    }
-                }
-
-                private void finish() {
-                    this.cancel();
-
-                    completableFuture.complete(null);
-
-                    new BukkitRunnable() {
-                        @Override
-                        public void run() {
-                            if (!IridiumSkyblock.getInstance().getConfiguration().generatorType.isTerrainGenerator()) {
-                                weWorld.fixLighting(selection.getChunks());
-                            }
-                        }
-                    }.runTask(IridiumSkyblock.getInstance());
-                }
-            }.runTaskTimer(IridiumSkyblock.getInstance(), delay, 1L);
-        }
-    }
-
-    public void deleteFAWE(World world, Location pos1, Location pos2, int minHeight, int maxHeight, CompletableFuture<Void> completableFuture, int delay) {
-
-        com.sk89q.worldedit.world.World weWorld = BukkitAdapter.adapt(world);
-
-        try (EditSession editSession = WorldEdit.getInstance().newEditSession(weWorld)) {
-            BlockVector3 min = BlockVector3.at(pos1.getBlockX(), minHeight, pos1.getBlockZ());
-            BlockVector3 max = BlockVector3.at(pos2.getBlockX(), maxHeight, pos2.getBlockZ());
-            CuboidRegion selection = new CuboidRegion(min, max);
-
-            com.sk89q.worldedit.world.block.BlockState air = BlockTypes.AIR.getDefaultState();
-            ExistingBlockMask nonAirMask = new ExistingBlockMask(editSession);
-
-            editSession.getChangeSet().setRecordChanges(false);
-            editSession.setReorderMode(EditSession.ReorderMode.FAST);
-            editSession.setSideEffectApplier(SideEffectSet.none());
-
-            editSession.replaceBlocks(selection, nonAirMask, (Pattern) air); // no style points :(
-
-            new BukkitRunnable() {
-                @Override
-                public void run() {
-                    if (!IridiumSkyblock.getInstance().getConfiguration().generatorType.isTerrainGenerator()) {
-                        weWorld.fixLighting(selection.getChunks());
-                    }
-                }
-            }.runTask(IridiumSkyblock.getInstance());
-
-        } catch (WorldEditException e) {
-            throw new RuntimeException(e);
-        }
-
-        completableFuture.complete(null);
-    }
-
-    // TODO: should consider chunk-batching like w/ WE
-    public void regenIA(World world, World regenWorld, Location pos1, Location pos2, int minHeight, int maxHeight, CompletableFuture<Void> completableFuture, int delay) {
-
-        Set<Long> modifiedChunks = new HashSet<>();
-
-        int minX = Math.min(pos1.getBlockX(), pos2.getBlockX());
-        int maxX = Math.max(pos1.getBlockX(), pos2.getBlockX());
-        int minZ = Math.min(pos1.getBlockZ(), pos2.getBlockZ());
-        int maxZ = Math.max(pos1.getBlockZ(), pos2.getBlockZ());
-
-        new BukkitRunnable() {
-
-            int y = minHeight;
-            int x = minX;
-            int z = minZ;
-
-            @Override
-            public void run() {
-                for (int i = 0; i < IridiumSkyblock.getInstance().getConfiguration().pasterLimitPerTick; i++) {
-
-                    Block blockA = regenWorld.getBlockAt(x, y, z);
-                    Block blockB = world.getBlockAt(x, y, z);
-
-                    while (blockA.isEmpty() || blockA.getBlockData().equals(blockB.getBlockData())) {
-                        if (increment()) return;
-                        blockA = regenWorld.getBlockAt(x, y, z);
-                        blockB = world.getBlockAt(x, y, z);
-                    }
-
-                    blockB.setBlockData(blockA.getBlockData(), false);
-                    modifiedChunks.add(((long) (x >> 4) << 32) | ((z >> 4) & 0xFFFFFFFFL));
-
-                    increment();
-                }
-            }
-
-            public boolean increment() {
-                if (++z > maxZ) {
-                    z = minZ;
-                    if (++x > maxX) {
-                        x = minX;
-                        if (++y > maxHeight) {
-                            finish();
-                            return true;
-                        }
-                    }
-                }
-                return false;
-            }
-
-            private void finish() {
-                this.cancel();
-
-                completableFuture.complete(null);
-
-                new BukkitRunnable() {
-                    @Override
-                    public void run() {
-                        modifiedChunks.forEach(key -> {world.refreshChunk((int) (key >> 32), key.intValue());});
-                    }
-                }.runTaskAsynchronously(IridiumSkyblock.getInstance());
-            }
-        }.runTaskTimer(IridiumSkyblock.getInstance(), delay, 1L);
-    }
-
-    public void regenWE(World world, World regenWorld, Location pos1, Location pos2, Location pos3, Location pos4, int minHeight, int maxHeight, CompletableFuture<Void> completableFuture, int delay) {
-
-        com.sk89q.worldedit.world.World weWorld = BukkitAdapter.adapt(world);
-        com.sk89q.worldedit.world.World weRegenWorld = BukkitAdapter.adapt(regenWorld);
-
-        // there's gotta be a more succinct method to do this
-        BlockVector3 min = BlockVector3.at(pos1.getBlockX(), minHeight, pos1.getBlockZ());
-        BlockVector3 max = BlockVector3.at(pos2.getBlockX(), maxHeight, pos2.getBlockZ());
-        CuboidRegion selection = new CuboidRegion(min, max);
-
-        BlockVector3 min2 = BlockVector3.at(pos3.getBlockX(), minHeight, pos3.getBlockZ());
-        BlockVector3 max2 = BlockVector3.at(pos4.getBlockX(), maxHeight, pos4.getBlockZ());
-        CuboidRegion selection2 = new CuboidRegion(min2, max2);
-
-        BlockArrayClipboard clipboard = new BlockArrayClipboard(selection);
-
-        // copy
-        try (EditSession editSession = WorldEdit.getInstance().newEditSession(weRegenWorld)) {
-            editSession.getChangeSet().setRecordChanges(false);
-            editSession.setReorderMode(EditSession.ReorderMode.FAST);
-            editSession.setSideEffectApplier(SideEffectSet.none());
-            ExistingBlockMask nonAirMask = new ExistingBlockMask(editSession);
-            editSession.setMask(nonAirMask);
-
-            ForwardExtentCopy forwardExtentCopy = new ForwardExtentCopy(
-                    weRegenWorld,
-                    selection,
-                    clipboard,
-                    selection.getMinimumPoint()
-            );
-
-            Operations.complete(forwardExtentCopy);
-
-        } catch (WorldEditException e) {
-            throw new RuntimeException(e);
-        }
-
-        // paste
-        EditSession editSession = WorldEdit.getInstance().newEditSession(weWorld);
-        editSession.getChangeSet().setRecordChanges(false);
-        editSession.setReorderMode(EditSession.ReorderMode.FAST);
-        editSession.setSideEffectApplier(SideEffectSet.none());
-
-        BlockVector3 offset = selection2.getMinimumPoint().subtract(clipboard.getMinimumPoint());
-
-        Mask clipboardMask = new ExistingBlockMask(clipboard);
-        MaskingExtent maskedClipboard = new MaskingExtent(clipboard, clipboardMask);
-
-        // grab all points and remove air blocks from the selection
-        List<BlockVector3> points = new ArrayList<>();
-        for (BlockVector3 point : clipboard.getRegion()) {
-            if (maskedClipboard.getMask().test(point)) {
-                points.add(point);
-            }
-        }
-
-        // group points into chunks
-        Map<BlockVector2, List<BlockVector3>> chunkBatches = new LinkedHashMap<>();
-        for (BlockVector3 pt : points) {
-            BlockVector2 chunkPos = BlockVector2.at(pt.getX() >> 4, pt.getZ() >> 4);
-            chunkBatches.computeIfAbsent(chunkPos, k -> new ArrayList<>()).add(pt);
-        }
-
-        // chunk list
-        List<Map.Entry<BlockVector2, List<BlockVector3>>> batches = new ArrayList<>(chunkBatches.entrySet());
-
-        new BukkitRunnable() {
-            int chunkIndex = 0;
-            int blockInsideChunkIndex = 0;
-            final int blocksPerTick = IridiumSkyblock.getInstance().getConfiguration().pasterLimitPerTick;
-            private final com.sk89q.worldedit.world.block.BlockState air =
-                    com.sk89q.worldedit.world.block.BlockTypes.AIR.getDefaultState();
-
-            @Override
-            public void run() {
-                int processedThisTick = 0;
-
-                while (processedThisTick < blocksPerTick) {
-
-                    if (chunkIndex >= batches.size()) {
-                        finish();
-                        return;
-                    }
-
-                    List<BlockVector3> currentChunkBlocks = batches.get(chunkIndex).getValue();
-
-                    while (blockInsideChunkIndex < currentChunkBlocks.size() && processedThisTick < blocksPerTick) {
-
-                        BlockVector3 point = currentChunkBlocks.get(blockInsideChunkIndex);
-                        com.sk89q.worldedit.world.block.BlockState block = clipboard.getBlock(point);
-
-                        try {
-                            editSession.setBlock(point.add(offset), block);
-                        } catch (MaxChangedBlocksException e) {
-                            finish();
-                            return;
-                        }
-
-                        blockInsideChunkIndex++;
-                        processedThisTick++;
-                    }
-
-                    if (blockInsideChunkIndex >= currentChunkBlocks.size()) {
-                        chunkIndex++;
-                        blockInsideChunkIndex = 0;
-                        editSession.flushSession(); // style points
-                    }
-                }
-            }
-
-            private void finish() {
-                this.cancel();
-
-                completableFuture.complete(null);
-
-                new BukkitRunnable() {
-                    @Override
-                    public void run() {
-                        if (!IridiumSkyblock.getInstance().getConfiguration().generatorType.isTerrainGenerator()) {
-                            weWorld.fixLighting(selection.getChunks());
-                        }
-                    }
-                }.runTask(IridiumSkyblock.getInstance());
-            }
-        }.runTaskTimer(IridiumSkyblock.getInstance(), delay, 1L);
-    }
-
-    public void regenFAWE(World world, World regenWorld, Location pos1, Location pos2, Location pos3, Location pos4, int minHeight, int maxHeight, CompletableFuture<Void> completableFuture, int delay) {
-
-        com.sk89q.worldedit.world.World weWorld = BukkitAdapter.adapt(world);
-        com.sk89q.worldedit.world.World weRegenWorld = BukkitAdapter.adapt(regenWorld);
-
-        BlockVector3 min = BlockVector3.at(pos1.getBlockX(), minHeight, pos1.getBlockZ());
-        BlockVector3 max = BlockVector3.at(pos2.getBlockX(), maxHeight, pos2.getBlockZ());
-        CuboidRegion selection = new CuboidRegion(min, max);
-
-        BlockVector3 min2 = BlockVector3.at(pos3.getBlockX(), minHeight, pos3.getBlockZ());
-        BlockVector3 max2 = BlockVector3.at(pos4.getBlockX(), maxHeight, pos4.getBlockZ());
-        CuboidRegion selection2 = new CuboidRegion(min2, max2);
-
-        BlockArrayClipboard clipboard = new BlockArrayClipboard(selection);
-
-        // copy
-        try (EditSession editSession = WorldEdit.getInstance().newEditSession(weRegenWorld)) {
-            editSession.getChangeSet().setRecordChanges(false);
-            ExistingBlockMask nonAirMask = new ExistingBlockMask(editSession);
-            editSession.setMask(nonAirMask);
-
-            ForwardExtentCopy forwardExtentCopy = new ForwardExtentCopy(
-                    weRegenWorld,
-                    selection,
-                    clipboard,
-                    selection.getMinimumPoint()
-            );
-
-            Operations.complete(forwardExtentCopy);
-
-        } catch (WorldEditException e) {
-            throw new RuntimeException(e);
-        }
-
-        // paste
-        try (EditSession editSession = WorldEdit.getInstance().newEditSession(weWorld)) {
-            editSession.getChangeSet().setRecordChanges(false);
-
-            Operation operation = new ClipboardHolder(clipboard)
-                    .createPaste(editSession)
-                    .to(selection2.getMinimumPoint())
-                    .copyBiomes(true)
-                    .copyEntities(false)
-                    .build();
-
-            Operations.complete(operation);
-        } catch (WorldEditException e) {
-            throw new RuntimeException(e);
-        }
-
-        completableFuture.complete(null);
-        weWorld.fixLighting(selection.getChunks());
-    }
+}
 }
