+++ src/main/java/com/iridium/iridiumskyblock/managers/SchematicManager.java
@@ -1,5 +1,6 @@
 package com.iridium.iridiumskyblock.managers;
 
+import com.iridium.iridiumcore.multiversion.IridiumInventory;
 import com.iridium.iridiumcore.utils.StringUtils;
 import com.iridium.iridiumskyblock.IridiumSkyblock;
 import com.iridium.iridiumskyblock.configs.Schematics;
@@ -10,12 +11,33 @@ import com.iridium.iridiumskyblock.schematics.SchematicAsync;
 import com.iridium.iridiumskyblock.schematics.SchematicPaster;
 import com.iridium.iridiumskyblock.schematics.WorldEdit;
 import com.iridium.iridiumteams.database.TeamBank;
+import com.sk89q.worldedit.EditSession;
+import com.sk89q.worldedit.MaxChangedBlocksException;
+import com.sk89q.worldedit.WorldEditException;
+import com.sk89q.worldedit.bukkit.BukkitAdapter;
+import com.sk89q.worldedit.extent.MaskingExtent;
+import com.sk89q.worldedit.extent.clipboard.BlockArrayClipboard;
+import com.sk89q.worldedit.function.mask.ExistingBlockMask;
+import com.sk89q.worldedit.function.mask.Mask;
+import com.sk89q.worldedit.function.operation.ForwardExtentCopy;
+import com.sk89q.worldedit.function.operation.Operation;
+import com.sk89q.worldedit.function.operation.Operations;
+import com.sk89q.worldedit.function.pattern.Pattern;
+import com.sk89q.worldedit.math.BlockVector2;
+import com.sk89q.worldedit.math.BlockVector3;
+import com.sk89q.worldedit.regions.CuboidRegion;
+import com.sk89q.worldedit.session.ClipboardHolder;
+import com.sk89q.worldedit.util.SideEffectSet;
+import com.sk89q.worldedit.world.block.BlockTypes;
 import net.milkbowl.vault.economy.Economy;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.World;
+import org.bukkit.block.Block;
 import org.bukkit.entity.Player;
+import org.bukkit.inventory.InventoryHolder;
+import org.bukkit.scheduler.BukkitRunnable;
 
 import java.io.File;
 import java.util.*;
@@ -175,4 +197,22 @@ public class SchematicManager {
         return String.valueOf(value);
     }
 
+    // IslandManager calls deleteIsland() and regenIsland() for each world.
+    // This is intentional, as we'll be reworking the regen command to only regenerate a specific dimension.
+
+    public CompletableFuture<Void> deleteIsland(World world, Location pos1, Location pos2, int minHeight, int maxHeight, CompletableFuture<Void> completableFuture, int delay) {
+
+        Bukkit.getScheduler().runTask(IridiumSkyblock.getInstance(), () -> {
+            schematicPaster.deleteIsland(world, pos1, pos2, minHeight, maxHeight, completableFuture, delay);
+        });
+        return completableFuture;
+    }
+
+    public CompletableFuture<Void> regenIsland(World world, World regenWorld, Location pos1, Location pos2, Location pos3, Location pos4, int minHeight, int maxHeight, CompletableFuture<Void> completableFuture, int delay) {
+
+        Bukkit.getScheduler().runTask(IridiumSkyblock.getInstance(), () -> {
+            schematicPaster.regenIsland(world, regenWorld, pos1, pos2, pos3, pos4, minHeight, maxHeight, completableFuture, delay);
+        });
+        return completableFuture;
+    }
 }
